Scripts/Bubble폴더를 정리한다
자원 버블은 Resource폴더로, 스페셜 버블은 Special폴더로 구분한다.
