+++ Assets/Scripts/Bubble/Resource.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab0e5795892b43149bba471ce72f2957
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
