+++ Assets/Scripts/Bubble/Resource/FoodBubble.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace InGame.Bubble
+{
+    public class FoodBubble : Bubble, IBubble
+    {
+        public void _BubbleClicked()
+        {
+            IncreaseResourceInBubble(Vehicles.CULTIVATION);
+        }
+    }
+
+}
