+++ Assets/Scripts/Bubble/Special.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae748cf17fac91f46bb241e5a474c67c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
